test: move "rendering" tests into dedicated file
This declutters the main `runtests` file a bit and makes it easier to
test only this particular piece of functionality (and to ignore it while
testing other pieces of the package).
